feat: restructure and enhance Remix and React Router integration with new plugins and request handlers
