+++ packages/remix-fastify/src/index.ts
@@ -2,5 +2,5 @@ export type {
   GetLoadContextFunction,
   RequestHandler,
   RemixFastifyOptions,
-} from "./remix";
-export { createRequestHandler, remixFastify } from "./remix";
+} from "./remix/index.ts";
+export { createRequestHandler, remixFastify } from "./remix/index.ts";
