+++ packages/remix-fastify/src/react-router/server.ts
@@ -1,29 +1,28 @@
-import type { AppLoadContext, ServerBuild } from "@remix-run/node";
-import {
-  createRequestHandler,
-  createReadableStreamFromReadable,
-} from "@remix-run/node";
+import { createReadableStreamFromReadable } from "@react-router/node";
 import type {
   FastifyRequest,
   FastifyReply,
   RouteGenericInterface,
 } from "fastify";
+import type { AppLoadContext, ServerBuild } from "react-router";
+import { createRequestHandler } from "react-router";
 
-import { createRequest, sendResponse } from "../shared";
+import { createRequest, sendResponse } from "../shared-server.ts";
 import type {
   GetLoadContextFunction as GenericGetLoadContextFunction,
   HttpServer,
   RequestHandler,
-} from "../shared";
+} from "../shared-server.ts";
 
-export type CreateRequestHandlerFunction = typeof createRemixRequestHandler;
+export type CreateRequestHandlerFunction =
+  typeof createReactRouterRequestHandler;
 export type GetLoadContextFunction<Server extends HttpServer = HttpServer> =
   GenericGetLoadContextFunction<Server, AppLoadContext>;
 
 /**
- * Returns a request handler for Fastify that serves the response using Remix.
+ * Returns a request handler for Fastify that serves the response using React Router.
  */
-export function createRemixRequestHandler<Server extends HttpServer>({
+export function createReactRouterRequestHandler<Server extends HttpServer>({
   build,
   getLoadContext,
   mode = process.env.NODE_ENV,
@@ -35,14 +34,14 @@ export function createRemixRequestHandler<Server extends HttpServer>({
   let handleRequest = createRequestHandler(build, mode);
 
   return async (request, reply) => {
-    let remixRequest = createRemixRequest(request, reply);
+    let reactRouterRequest = createReactRouterRequest(request, reply);
     let loadContext = await getLoadContext?.(request, reply);
-    let response = await handleRequest(remixRequest, loadContext);
+    let response = await handleRequest(reactRouterRequest, loadContext);
     return sendResponse(reply, response);
   };
 }
 
-export function createRemixRequest<Server extends HttpServer>(
+export function createReactRouterRequest<Server extends HttpServer>(
   request: FastifyRequest<RouteGenericInterface, Server>,
   reply: FastifyReply<RouteGenericInterface, Server>,
 ): Request {
