+++ packages/remix-fastify/tsconfig.json
@@ -5,6 +5,7 @@
   "compilerOptions": {
     "forceConsistentCasingInFileNames": true,
     "outDir": "./dist",
-    "rootDir": "."
+    "rootDir": ".",
+    "allowImportingTsExtensions": true
   }
 }
