+++ packages/remix-fastify/tsdown.config.ts
@@ -3,8 +3,8 @@ import pkg from "./package.json" with { type: "json" };
 
 export default defineConfig({
   entry: {
-    index: "./src/remix.ts",
-    "react-router": "./src/react-router.ts",
+    index: "./src/remix/index.ts",
+    "react-router": "./src/react-router/index.ts",
   },
   exports: true,
   sourcemap: true,
